fixed future gameplay bug where if someone shot 500 bullets then switched to a nuke they would apply 500 nukes worth of damage

+++ src/shield/CapitalShipShield.java
@@ -1,7 +1,7 @@
 package shield;
 
 public class CapitalShipShield extends ship.Shield {
-     private double POWER = 5000;
+     private int POWER = 5000;
     
     public CapitalShipShield() {
         setPower(POWER);
