+++ src/shield/PlayerShield.java
@@ -1,7 +1,7 @@
 package shield;
 
 public class PlayerShield extends ship.Shield {
-    final private double POWER = 1000;
+    final private int POWER = 1000;
     
     public PlayerShield() {
         setPower(POWER);
